+++ src/OpenTelemetry.AutoInstrumentation/Configurations/LogSettings.cs
@@ -29,11 +29,6 @@ internal class LogSettings : Settings
     public bool IncludeFormattedMessage { get; private set; }
 
     /// <summary>
-    /// Gets a value indicating whether the console exporter is enabled.
-    /// </summary>
-    public bool ConsoleExporterEnabled { get; private set; }
-
-    /// <summary>
     /// Gets the list of enabled instrumentations.
     /// </summary>
     public IReadOnlyList<LogInstrumentation> EnabledInstrumentations { get; private set; } = new List<LogInstrumentation>();
@@ -43,16 +38,21 @@ internal class LogSettings : Settings
     /// </summary>
     public OtlpSettings? OtlpSettings { get; private set; }
 
+    /// <summary>
+    /// Gets or sets a value indicating whether the console exporter is enabled.
+    /// </summary>
+    private bool ConsoleExporterEnabled { get; set; }
+
     protected override void OnLoad(Configuration configuration)
     {
         LogsEnabled = configuration.GetBool(ConfigurationKeys.Logs.LogsEnabled) ?? true;
+        ConsoleExporterEnabled = configuration.GetBool(ConfigurationKeys.Logs.ConsoleExporterEnabled) ?? false;
         LogExporters = ParseLogExporter(configuration);
         if (LogExporters.Contains(LogExporter.Otlp))
         {
             OtlpSettings = new OtlpSettings(OtlpSignalType.Logs, configuration);
         }
 
-        ConsoleExporterEnabled = configuration.GetBool(ConfigurationKeys.Logs.ConsoleExporterEnabled) ?? false;
         IncludeFormattedMessage = configuration.GetBool(ConfigurationKeys.Logs.IncludeFormattedMessage) ?? false;
 
         var instrumentationEnabledByDefault =
@@ -64,7 +64,7 @@ internal class LogSettings : Settings
             enabledConfigurationTemplate: ConfigurationKeys.Logs.EnabledLogsInstrumentationTemplate);
     }
 
-    private static IReadOnlyList<LogExporter> ParseLogExporter(Configuration configuration)
+    private IReadOnlyList<LogExporter> ParseLogExporter(Configuration configuration)
     {
         var logExporterEnvVar = configuration.GetString(ConfigurationKeys.Logs.Exporter);
 
@@ -83,6 +83,9 @@ internal class LogSettings : Settings
                 case Constants.ConfigurationValues.Exporters.Otlp:
                     exporters.Add(LogExporter.Otlp);
                     break;
+                case Constants.ConfigurationValues.Exporters.Console:
+                    exporters.Add(LogExporter.Console);
+                    break;
                 case Constants.ConfigurationValues.None:
                     break;
                 default:
@@ -96,6 +99,16 @@ internal class LogSettings : Settings
                     Logger.Error($"Log exporter '{exporterName}' is not supported.");
                     break;
             }
+        }
+
+        if (ConsoleExporterEnabled)
+        {
+            Logger.Warning($"The '{ConfigurationKeys.Logs.ConsoleExporterEnabled}' environment variable is deprecated and " +
+                "will be removed in the next minor release. " +
+                "Please set the console exporter using OTEL_LOGS_EXPORTER environmental variable. " +
+                "Refer to the updated documentation for details.");
+
+            exporters.Add(LogExporter.Console);
         }
 
         return exporters;