+++ src/OpenTelemetry.AutoInstrumentation/Configurations/MetricSettings.cs
@@ -24,11 +24,6 @@ internal class MetricSettings : Settings
     public IReadOnlyList<MetricsExporter> MetricExporters { get; private set; } = new List<MetricsExporter>();
 
     /// <summary>
-    /// Gets a value indicating whether the console exporter is enabled.
-    /// </summary>
-    public bool ConsoleExporterEnabled { get; private set; }
-
-    /// <summary>
     /// Gets the list of enabled meters.
     /// </summary>
     public IReadOnlyList<MetricInstrumentation> EnabledInstrumentations { get; private set; } = new List<MetricInstrumentation>();
@@ -42,6 +37,11 @@ internal class MetricSettings : Settings
     /// Gets metrics OTLP Settings.
     /// </summary>
     public OtlpSettings? OtlpSettings { get; private set; }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the console exporter is enabled.
+    /// </summary>
+    private bool ConsoleExporterEnabled { get; set; }
 
     protected override void OnLoad(Configuration configuration)
     {
@@ -73,7 +73,7 @@ internal class MetricSettings : Settings
         MetricsEnabled = configuration.GetBool(ConfigurationKeys.Metrics.MetricsEnabled) ?? true;
     }
 
-    private static IReadOnlyList<MetricsExporter> ParseMetricExporter(Configuration configuration)
+    private IReadOnlyList<MetricsExporter> ParseMetricExporter(Configuration configuration)
     {
         var metricsExporterEnvVar = configuration.GetString(ConfigurationKeys.Metrics.Exporter);
 
@@ -95,9 +95,11 @@ internal class MetricSettings : Settings
                 case Constants.ConfigurationValues.Exporters.Prometheus:
                     exporters.Add(MetricsExporter.Prometheus);
                     break;
+                case Constants.ConfigurationValues.Exporters.Console:
+                    exporters.Add(MetricsExporter.Console);
+                    break;
                 case Constants.ConfigurationValues.None:
                     break;
-
                 default:
                     if (configuration.FailFast)
                     {
@@ -109,6 +111,16 @@ internal class MetricSettings : Settings
                     Logger.Error($"Metric exporter '{exporterName}' is not supported.");
                     break;
             }
+        }
+
+        if (ConsoleExporterEnabled)
+        {
+            Logger.Warning($"The '{ConfigurationKeys.Metrics.ConsoleExporterEnabled}' environment variable is deprecated and " +
+                "will be removed in the next minor release. " +
+                "Please set the console exporter using OTEL_METRICS_EXPORTER environmental variable. " +
+                "Refer to the updated documentation for details.");
+
+            exporters.Add(MetricsExporter.Console);
         }
 
         return exporters;