+++ src/OpenTelemetry.AutoInstrumentation/Configurations/MetricsExporter.cs
@@ -17,4 +17,9 @@ internal enum MetricsExporter
     /// Prometheus exporter.
     /// </summary>
     Prometheus,
+
+    /// <summary>
+    /// Console exporter.
+    /// </summary>
+    Console,
 }