+++ src/OpenTelemetry.AutoInstrumentation/Configurations/TracerSettings.cs
@@ -29,11 +29,6 @@ internal class TracerSettings : Settings
     public IReadOnlyList<TracesExporter> TracesExporters { get; private set; } = new List<TracesExporter>();
 
     /// <summary>
-    /// Gets a value indicating whether the console exporter is enabled.
-    /// </summary>
-    public bool ConsoleExporterEnabled { get; private set; }
-
-    /// <summary>
     /// Gets the list of enabled instrumentations.
     /// </summary>
     public IReadOnlyList<TracerInstrumentation> EnabledInstrumentations { get; private set; } = new List<TracerInstrumentation>();
@@ -57,6 +52,11 @@ internal class TracerSettings : Settings
     /// Gets tracing OTLP Settings.
     /// </summary>
     public OtlpSettings? OtlpSettings { get; private set; }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the console exporter is enabled.
+    /// </summary>
+    private bool ConsoleExporterEnabled { get; set; }
 
     protected override void OnLoad(Configuration configuration)
     {
@@ -100,7 +100,7 @@ internal class TracerSettings : Settings
         InstrumentationOptions = new InstrumentationOptions(configuration);
     }
 
-    private static IReadOnlyList<TracesExporter> ParseTracesExporter(Configuration configuration)
+    private IReadOnlyList<TracesExporter> ParseTracesExporter(Configuration configuration)
     {
         var tracesExporterEnvVar = configuration.GetString(ConfigurationKeys.Traces.Exporter);
 
@@ -125,6 +125,9 @@ internal class TracerSettings : Settings
                     break;
                 case Constants.ConfigurationValues.None:
                     break;
+                case Constants.ConfigurationValues.Exporters.Console:
+                    exporters.Add(TracesExporter.Console);
+                    break;
                 default:
                     if (configuration.FailFast)
                     {
@@ -136,6 +139,16 @@ internal class TracerSettings : Settings
                     Logger.Error($"Traces exporter '{exporterName}' is not supported.");
                     break;
             }
+        }
+
+        if (ConsoleExporterEnabled)
+        {
+            Logger.Warning($"The '{ConfigurationKeys.Traces.ConsoleExporterEnabled}' environment variable is deprecated and " +
+                "will be removed in the next minor release. " +
+                "Please set the console exporter using OTEL_TRACES_EXPORTER environmental variable. " +
+                "Refer to the updated documentation for details.");
+
+            exporters.Add(TracesExporter.Console);
         }
 
         return exporters;