+++ src/OpenTelemetry.AutoInstrumentation/Instrumentations/StackExchangeRedis/StackExchangeRedisIntegrationAsync.cs
@@ -4,33 +4,12 @@
 #if NET6_0_OR_GREATER
 
 using OpenTelemetry.AutoInstrumentation.CallTarget;
-using OpenTelemetry.AutoInstrumentation.Configurations;
 
 namespace OpenTelemetry.AutoInstrumentation.Instrumentations.StackExchangeRedis;
 
 /// <summary>
 /// StackExchange.Redis.ConnectionMultiplexer calltarget instrumentation
 /// </summary>
-[InstrumentMethod(// releases 2.0.495 - 2.1.39
-    assemblyName: StackExchangeRedisConstants.AssemblyName,
-    typeName: StackExchangeRedisConstants.ConnectionMultiplexerTypeName,
-    methodName: StackExchangeRedisConstants.ConnectImplAsyncMethodName,
-    returnTypeName: StackExchangeRedisConstants.TaskConnectionMultiplexerTypeName,
-    parameterTypeNames: new[] { ClrNames.Object, StackExchangeRedisConstants.TextWriterTypeName },
-    minimumVersion: StackExchangeRedisConstants.MinimumVersion,
-    maximumVersion: StackExchangeRedisConstants.MaximumVersion,
-    integrationName: StackExchangeRedisConstants.IntegrationName,
-    type: InstrumentationType.Trace)]
-[InstrumentMethod(// releases 2.1.50 - 2.5.43
-    assemblyName: StackExchangeRedisConstants.AssemblyName,
-    typeName: StackExchangeRedisConstants.ConnectionMultiplexerTypeName,
-    methodName: StackExchangeRedisConstants.ConnectImplAsyncMethodName,
-    returnTypeName: StackExchangeRedisConstants.TaskConnectionMultiplexerTypeName,
-    parameterTypeNames: new[] { StackExchangeRedisConstants.ConfigurationOptionsTypeName, StackExchangeRedisConstants.TextWriterTypeName },
-    minimumVersion: StackExchangeRedisConstants.MinimumVersion,
-    maximumVersion: StackExchangeRedisConstants.MaximumVersion,
-    integrationName: StackExchangeRedisConstants.IntegrationName,
-    type: InstrumentationType.Trace)]
 [InstrumentMethod(// releases 2.5.61+
     assemblyName: StackExchangeRedisConstants.AssemblyName,
     typeName: StackExchangeRedisConstants.ConnectionMultiplexerTypeName,