+++ CHANGELOG.md
@@ -12,6 +12,12 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 - Support for Operating System resource detector.
 - Added support for OTEL_TRACES_EXPORTER, OTEL_METRICS_EXPORTER, OTEL_LOGS_EXPORTER
   to handle comma-separated list.
+- The environment variables `OTEL_TRACES_EXPORTER`, `OTEL_METRICS_EXPORTER`,
+  and `OTEL_LOGS_EXPORTER` now support configuring console exporters for traces,
+  metrics, and logs, respectively.
+- Environment variables `OTEL_DOTNET_AUTO_TRACES_CONSOLE_EXPORTER_ENABLED`,
+  `OTEL_DOTNET_AUTO_METRICS_CONSOLE_EXPORTER_ENABLED`, and
+  `OTEL_DOTNET_AUTO_LOGS_CONSOLE_EXPORTER_ENABLED` are now marked as deprecated.
 - Support signal specific OTLP exporter variables (See [docs](/docs/config.md#otlp)):
   - `OTEL_EXPORTER_OTLP_TRACES_ENDPOINT`,
   - `OTEL_EXPORTER_OTLP_TRACES_HEADERS`,
@@ -28,11 +34,16 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Changed
 
+- Referencing `OpenTelemetry.AutoInstrumentation` manually no longer visibly injects
+  instrumentation scripts into projects in an editor's solution window.
+
 #### Dependency updates
 
 - Following packages updated
   - `OpenTelemetry.Exporter.Prometheus.HttpListener` from `1.9.0-beta.1` to `1.9.0-beta.2`,
   - `OpenTelemetry.Shims.OpenTracing` from `1.9.0-beta.1` to `1.9.0-beta.2`.
+- .NET only, following packages updated
+  - `OpenTelemetry.Instrumentation.StackExchangeRedis` from `1.0.0-rc9.15` to `1.9.0-beta.1`.
 - .NET Framework only, following packages updated
   - `Google.Protobuf` updated from `3.27.1` to `3.27.2`,
   - `Grpc.Core.Api` from `2.63.0` to `2.64.0`,
@@ -45,6 +56,7 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 - Support for macOS Big Sur 11 x64.
   macOs libraries are built and tested against [macOS Monterey 12 x64](https://github.com/actions/runner-images/blob/main/images/macos/macos-12-Readme.md).
+- Support for StackExchange.Redis < 2.6.122.
 
 ### Fixed